New inventory slot
Working with an engine for the first time in my life: day two. I managed to add a slot for backpacks (!!!) to the game. It is fully stable and adopts the ability to additionally edit the carrying weight for backpacks, which will be their main (and most likely only) feature :)