Now only bin, no binaries
I used to duplicate the contents of the "binaries" folder into the "bin" folder to make it easier for users to copy files into the game files. Now all new files are created directly into the "bin" folder, which allows me to remove the now unnecessary "binaries" folder. The size of the overall archive will also become much smaller